refs(mail): Move `should_notify` to `MailAdapter`.
This copies the logic for `should_notify` to the `MailAdapter` class, and has
`MailPlugin.should_notify` use it instead.

Note that we removed the code to handle digests being disabled - This is hardcoded to `True` in the
`digests.RedisBackend`, and so we don't need to handle the separate logic here, since we'll always
have digests available.